refactor(strategy): 重构策略相关代码
- 修改策略类型字段名称
- 更新策略评估和告警逻辑
- 优化策略构建器和数据源处理
- 调整策略相关模型和BO

+++ cmd/server/palace/internal/biz/bo/strategy.go
@@ -160,7 +160,7 @@ type (
 		// 告警页面
 		AlarmPageIds []uint32 `json:"alarmPageIds"`
 		// 响应时间 s
-		ResponseTime uint32 `json:"responseTime"`
+		ResponseTime float64 `json:"responseTime"`
 		// 状态码
 		StatusCode string `json:"statusCodes"`
 		// 请求体
